Ignore deleted functions in the cleanup script

+++ docs/examples/scripts/cleanup.sh
@@ -2,17 +2,21 @@
 
 . $(dirname "$0")/common.sh
 
-aws lambda delete-function --no-cli-pager \
-  --function-name "${BUFFERED_FUNCTION_NAME}" ||
-    _panic "Can't delete the ${BUFFERED_FUNCTION_NAME} function"
+if (aws lambda get-function --function-name "${BUFFERED_FUNCTION_NAME}" --no-cli-pager > /dev/null 2> /dev/null); then
+  aws lambda delete-function --no-cli-pager \
+    --function-name "${BUFFERED_FUNCTION_NAME}" ||
+      _panic "Can't delete the ${BUFFERED_FUNCTION_NAME} function"
 
-echo "The ${BUFFERED_FUNCTION_NAME} function was deleted"
+  echo "The ${BUFFERED_FUNCTION_NAME} function was deleted"
+fi
 
-aws lambda delete-function --no-cli-pager \
-  --function-name "${REAL_TIME_FUNCTION_NAME}" ||
-    _panic "Can't delete the ${REAL_TIME_FUNCTION_NAME} function"
+if (aws lambda get-function --function-name "${REAL_TIME_FUNCTION_NAME}" --no-cli-pager > /dev/null 2> /dev/null); then
+  aws lambda delete-function --no-cli-pager \
+    --function-name "${REAL_TIME_FUNCTION_NAME}" ||
+      _panic "Can't delete the ${REAL_TIME_FUNCTION_NAME} function"
 
-echo "The ${REAL_TIME_FUNCTION_NAME} function was deleted"
+  echo "The ${REAL_TIME_FUNCTION_NAME} function was deleted"
+fi
 
 aws iam detach-role-policy --no-cli-pager \
   --role-name "${ROLE_NAME}" \
